+++ src/ios/LineLogin.swift
@@ -6,37 +6,56 @@ import LineSDK
     struct ErrorCode {
         static let ParameterError = -1
         static let SDKError = -2
+        static let UnknownError = -3
     }
 
     @objc func initialize(_ command: CDVInvokedUrlCommand) {
         
-        let params = command.arguments[0] as AnyObject
-        
-        guard let channelID = params["channel_id"] as? String else {
+        guard command.arguments.count > 0, let params = command.arguments[0] as? [String: Any] else {
             self.parameterError(command: command, description: "channel_id is required")
             return
         }
         
-        LoginManager.shared.setup(channelID: channelID, universalLinkURL: nil)
+        guard let channelID = params["channel_id"] as? String, !channelID.isEmpty else {
+            self.parameterError(command: command, description: "channel_id is required")
+            return
+        }
+        
+        if !LoginManager.shared.isSetupFinished {
+            LoginManager.shared.setup(channelID: channelID, universalLinkURL: nil)
+        }
         let result = CDVPluginResult(status: CDVCommandStatus_OK)
         commandDelegate.send(result, callbackId:command.callbackId)
     }
     
-    func _login(_ command: CDVInvokedUrlCommand, options: LoginManagerOptions) {
-        LoginManager.shared.login(permissions: [.profile, .openID, .email], in: self.viewController, options: options) {
+    func _login(_ command: CDVInvokedUrlCommand, onlyWebLogin: Bool) {
+        guard LoginManager.shared.isSetupFinished else {
+            self.parameterError(command: command, description: "initialize must be called before login")
+            return
+        }
+
+        var parameters = LoginManager.Parameters()
+        parameters.onlyWebLogin = onlyWebLogin
+        LoginManager.shared.login(permissions: [.profile, .openID, .email], in: self.viewController, parameters: parameters) {
             result in
             switch result {
             case .success(let loginResult):
-                var data = ["userID":nil, "displayName":nil, "pictureURL":nil, "email":nil] as [String : Any?]
+                var data = [String : Any]()
                 
-                data.updateValue(loginResult.userProfile?.displayName, forKey: "displayName")
-                data.updateValue(loginResult.userProfile?.userID, forKey: "userID")
-                data.updateValue(loginResult.accessToken.IDToken?.payload.email, forKey: "email")
+                if let displayName = loginResult.userProfile?.displayName {
+                    data["displayName"] = displayName
+                }
+                if let userID = loginResult.userProfile?.userID {
+                    data["userID"] = userID
+                }
+                if let email = loginResult.accessToken.IDToken?.payload.email {
+                    data["email"] = email
+                }
                 if let pictureURL = loginResult.userProfile?.pictureURL {
-                    data.updateValue(String(describing: pictureURL), forKey: "pictureURL")
+                    data["pictureURL"] = String(describing: pictureURL)
                 }
                 
-                let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data as [AnyHashable : Any])
+                let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data)
                 self.commandDelegate.send(result, callbackId:command.callbackId)
             case .failure(let error):
                 self.sdkError(command: command, error: error)
@@ -45,11 +64,11 @@ import LineSDK
     }
     
     @objc func loginWeb(_ command: CDVInvokedUrlCommand) {
-        self._login(command, options: LoginManagerOptions.init(rawValue: LoginManagerOptions.onlyWebLogin.rawValue))
+        self._login(command, onlyWebLogin: true)
     }
     
     @objc func login(_ command: CDVInvokedUrlCommand) {
-        self._login(command, options: LoginManagerOptions())
+        self._login(command, onlyWebLogin: false)
     }
     
     @objc func logout(_ command: CDVInvokedUrlCommand) {
@@ -67,19 +86,18 @@ import LineSDK
     @objc func getAccessToken(_ command: CDVInvokedUrlCommand) {
         
         guard let currentAccessToken = AccessTokenStore.shared.current else {
-            let result = CDVPluginResult(status: CDVCommandStatus_ERROR)
-            commandDelegate.send(result, callbackId:command.callbackId)
+            self.sdkError(command: command, sdkErrorCode: "NO_ACCESS_TOKEN", description: "No current LINE access token.")
             return
         }
         
-        let data = ["accessToken":currentAccessToken.value, "expireTime":currentAccessToken.expiresAt.timeIntervalSince1970] as [String : Any?]
-        let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data as [AnyHashable : Any])
+        let data = ["accessToken":currentAccessToken.value, "expireTime":currentAccessToken.expiresAt.timeIntervalSince1970] as [String : Any]
+        let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data)
         commandDelegate.send(result, callbackId:command.callbackId)
     }
     
     @objc func verifyAccessToken(_ command: CDVInvokedUrlCommand) {
         
-        API.verifyAccessToken { (result) in
+        API.Auth.verifyAccessToken { (result) in
             switch result {
             case .success( _):
                 let result = CDVPluginResult(status: CDVCommandStatus_OK)
@@ -92,11 +110,11 @@ import LineSDK
     
     @objc func refreshAccessToken(_ command: CDVInvokedUrlCommand) {
 
-        API.refreshAccessToken { (result) in
+        API.Auth.refreshAccessToken { (result) in
             switch result {
             case .success(let accessToken):
-                let data = ["accessToken":accessToken.value, "expireTime":accessToken.expiresAt.timeIntervalSince1970] as [String : Any?]
-                let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data as [AnyHashable : Any])
+                let data = ["accessToken":accessToken.value, "expireTime":accessToken.expiresAt.timeIntervalSince1970] as [String : Any]
+                let result = CDVPluginResult(status: CDVCommandStatus_OK, messageAs:data)
                 self.commandDelegate.send(result, callbackId:command.callbackId)
             case .failure(let error):
                 self.sdkError(command: command, error: error)
@@ -112,7 +130,11 @@ import LineSDK
     
     private func sdkError(command: CDVInvokedUrlCommand, error: LineSDKError) {
         let description = error.errorDescription ?? ""
-        let err = ["code":ErrorCode.SDKError, "sdkErrorCode":error.errorCode, "description": description] as [AnyHashable : Any]
+        self.sdkError(command: command, sdkErrorCode: String(error.errorCode), description: description)
+    }
+
+    private func sdkError(command: CDVInvokedUrlCommand, sdkErrorCode: String, description: String) {
+        let err = ["code":ErrorCode.SDKError, "sdkErrorCode":sdkErrorCode, "description": description] as [AnyHashable : Any]
         let result = CDVPluginResult(status: CDVCommandStatus_ERROR, messageAs: err)
         self.commandDelegate.send(result, callbackId:command.callbackId)
     }